backend/requirements.txt
flask
librosa
transformers
diffusers
torch
opencv-python
requests
streamlit
soundfile
ffmpeg-python
pydub
openai-whisper